requirements.txt
Django==1.6.6
Pillow==2.5.3
South==1.0
argparse==1.2.1
captcha==0.1.1
django-notifier==0.7
django-registration==1.0
django-simple-captcha==0.4.2
raven==5.1.1
six==1.8.0
wsgiref==0.1.2
